refactor(models): one schema for the order, shared by the site and its test
The browser test checked a shape of its own, so it could have accepted a file the site refuses. The schema now lives apart from the file it validates, and both read it.
