+++ apps/website/src/config/workshop-model-order.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import { workshopModelOrderSchema } from './workshop-model-order'
+import { workshopModelOrderSchema } from './workshop-model-order.schema'
 
 const order: unknown = {
   measuredOn: '2026-09-12',
