+++ apps/website/src/config/workshop-model-order.ts
@@ -1,19 +1,5 @@
-import { z } from 'astro/zod'
-
 import orderJson from '../content/workshop-model-order.json'
-
-export const workshopModelOrderSchema = z.object({
-  measuredOn: z.string(),
-  windowDays: z.number().int().positive(),
-  note: z.string(),
-  // A slug listed twice would quietly take its last position and rank one model
-  // wrong, with nothing to show for it.
-  slugs: z
-    .array(z.string())
-    .refine((slugs) => new Set(slugs).size === slugs.length, {
-      message: 'workshop-model-order.json lists a slug more than once'
-    })
-})
+import { workshopModelOrderSchema } from './workshop-model-order.schema'
 
 const order = workshopModelOrderSchema.parse(orderJson)
